+++ src/context.ts
@@ -7,6 +7,10 @@ import { EventEmitter } from 'events';
 
 import * as winston from 'winston'
 
+export interface Reference extends Releasable {
+	apply(...args: any[]): Promise<any>
+}
+
 export interface Releasable {
 	release(): void
 }
@@ -38,8 +42,8 @@ export class Context extends EventEmitter {
 
 	timeouts: { [id: number]: NodeJS.Timer }
 	intervals: { [id: number]: NodeJS.Timer }
-	callbacks: ivm.Reference<Function>[]
-	fireFetchEventFn?: ivm.Reference<Function>
+	callbacks: Reference[]
+	fireFetchEventFn?: Reference
 
 	releasables: Releasable[]
 
@@ -63,14 +67,14 @@ export class Context extends EventEmitter {
 		this.logMetadata = {}
 	}
 
-	addCallback(fn: ivm.Reference<Function>) {
+	addCallback(fn: Reference) {
 		this.addReleasable(fn)
 		this.callbacks.push(fn)
 		log.silly("Added callback", fn)
 		this.emit("callbackAdded", fn)
 	}
 
-	async applyCallback(fn: ivm.Reference<Function>, args: any[], opts?: any) {
+	async applyCallback(fn: Reference, args: any[], opts?: any) {
 		try {
 			return await this._applyCallback(fn, args, opts)
 		} catch (err) {
@@ -79,7 +83,7 @@ export class Context extends EventEmitter {
 		}
 	}
 
-	async _applyCallback(fn: ivm.Reference<Function>, args: any[], opts?: any) {
+	async _applyCallback(fn: Reference, args: any[], opts?: any) {
 		log.silly("Applying callback", fn, args)
 		try {
 			if (this.iso.isDisposed)
@@ -103,7 +107,7 @@ export class Context extends EventEmitter {
 		}
 	}
 
-	async tryCallback(fn: ivm.Reference<Function>, args: any[], opts?: any) {
+	async tryCallback(fn: Reference, args: any[], opts?: any) {
 		try {
 			return await this._applyCallback(fn, args, opts)
 		} catch (err) {
@@ -111,7 +115,7 @@ export class Context extends EventEmitter {
 		}
 	}
 
-	log(lvl: string, msg: string, meta?: any, cb?: ivm.Reference<Function>) {
+	log(lvl: string, msg: string, meta?: any, cb?: Reference) {
 		if (cb)
 			this.addCallback(cb)
 		this.logger.log(lvl, msg, Object.assign({}, this.persistentLogMetadata, this.logMetadata, meta || {}), (error?: any, level?: string, msg?: string, meta?: any) => {
@@ -120,15 +124,15 @@ export class Context extends EventEmitter {
 		})
 	}
 
-	addReleasable(ref: ivm.Reference<any>): ivm.Reference<any>;
+	addReleasable(ref: Reference): Reference;
 	addReleasable(ec: ivm.ExternalCopy<any>): ivm.ExternalCopy<any>;
-	addReleasable(rel: ivm.Reference<any> | ivm.ExternalCopy<any>) {
+	addReleasable(rel: Reference | ivm.ExternalCopy<any>) {
 		if (this.releasables.indexOf(rel) === -1)
 			this.releasables.push(rel)
 		return rel
 	}
 
-	setTimeout(fn: ivm.Reference<Function>, timeout: number) {
+	setTimeout(fn: Reference, timeout: number) {
 		const id = ++this.currentTimerId
 		this.timeouts[id] = setTimeout(() => { this.applyCallback(fn, []) }, timeout)
 		this.addCallback(fn)
@@ -141,7 +145,7 @@ export class Context extends EventEmitter {
 		return
 	}
 
-	setInterval(fn: ivm.Reference<Function>, every: number) {
+	setInterval(fn: Reference, every: number) {
 		const id = ++this.currentTimerId
 		// we don't add interval callbacks because we will clear them at the very end
 		this.intervals[id] = setInterval(() => { fn.apply(null, []) }, every)
@@ -159,31 +163,8 @@ export class Context extends EventEmitter {
 		await Promise.all([
 			this.set('global', this.global.derefInto()),
 			this.set('_ivm', ivm),
-			// this.set('_setTimeout', new ivm.Reference((fn: ivm.Reference<Function>, timeout: number): number => {
-			// 	const id = ++this.currentTimerId
-			// 	this.timeouts[id] = setTimeout(() => { this.applyCallback(fn, []) }, timeout)
-			// 	this.addCallback(fn)
-			// 	return id
-			// })),
-			// this.set('_clearTimeout', new ivm.Reference((id: number): void => {
-			// 	clearTimeout(this.timeouts[id])
-			// 	delete this.timeouts[id]
-			// 	return
-			// })),
-			// this.set('_setInterval', new ivm.Reference((fn: ivm.Reference<Function>, timeout: number): number => {
-			// 	const id = ++this.currentTimerId
-			// 	// we don't add interval callbacks because we will clear them at the very end
-			// 	this.intervals[id] = setInterval(() => { fn.apply(null, []) }, timeout)
-			// 	this.addReleasable(fn)
-			// 	return id
-			// })),
-			// this.set('_clearInterval', new ivm.Reference((id: number): void => {
-			// 	clearInterval(this.intervals[id])
-			// 	delete this.intervals[id]
-			// 	return
-			// })),
 			this.set("_dispatch", new ivm.Reference((name: string, ...args: any[]) => {
-				bridge.dispatch(this, name, ...args)
+				return bridge.dispatch(this, name, ...args)
 			})),
 			this.set('_log', new ivm.Reference(function (lvl: string, ...args: any[]) {
 				log.log(lvl, args[0], ...args.slice(1))
@@ -194,7 +175,7 @@ export class Context extends EventEmitter {
 		try {
 			await bootstrapFn.apply()
 		} finally {
-			bootstrapFn.release()
+			tryRelease(bootstrapFn)
 		}
 
 		return
@@ -253,20 +234,26 @@ export class Context extends EventEmitter {
 	}
 
 	async release() {
-		try {
-			const teardownFn = await this.global.get("teardown")
-			await teardownFn.apply(null, [])
-			teardownFn.release()
-			this.fireFetchEventFn && this.fireFetchEventFn.release()
-			this.global.release()
-			this.callbacks = []
-			this.intervals = {}
-			this.timeouts = {}
-			this.logger.close()
-			this.ctx.release()
-		} catch (err) {
-			log.error("error releasing context:", err.stack)
+		await this.runV8Teardown()
+		this.fireFetchEventFn && tryRelease(this.fireFetchEventFn)
+		tryRelease(this.global)
+		this.callbacks = []
+		this.intervals = {}
+		this.timeouts = {}
+		tryRelease(this.ctx)
+	}
+
+	// forcibly destroy everything about this context
+	destroy() {
+		this.clearIntervals()
+		this.clearTimeouts()
+
+		for (let cb of this.callbacks) {
+			this.tryCallback(cb, ["Destroying context."])
 		}
+
+		this.releaseAll()
+		this.release()
 	}
 
 	async finalize() {
@@ -288,27 +275,67 @@ export class Context extends EventEmitter {
 				this.on("callbackApplied", cbFn)
 			})
 		} finally {
+
+			await this.runV8Finalize()
+
 			// clear all intervals no matter what
-			for (const [id, t] of Object.entries(this.intervals)) {
-				clearInterval(t)
-				delete this.intervals[parseInt(id)] // stupid ts.
-			}
-			let rel;
-			while (rel = this.releasables.pop()) {
-				try {
-					rel.release()
-				} catch (e) {
-					// console.error("RELEASE ERROR:", e.stack)
-					// don't really care
-				}
-			}
+			this.clearIntervals()
+			this.releaseAll()
 			this.logMetadata = {} // reset log meta data!
 		}
 	}
+
+	async runV8Finalize() {
+		try {
+			const finalizeFn = await this.global.get("finalize")
+			await finalizeFn.apply(null, [])
+			tryRelease(finalizeFn)
+		} catch (e) {
+
+		}
+	}
+
+	async runV8Teardown() {
+		try {
+			const teardownFn = await this.global.get("teardown")
+			await teardownFn.apply(null, [])
+			tryRelease(teardownFn)
+		} catch (err) {
+			log.error("error tearing down v8:", err.stack)
+		}
+	}
+
+	releaseAll() {
+		let rel;
+		while (rel = this.releasables.pop()) {
+			tryRelease(rel)
+		}
+	}
+
+	clearTimeouts() {
+		for (const [id, t] of Object.entries(this.timeouts)) {
+			clearTimeout(t)
+			delete this.timeouts[parseInt(id, 10)]
+		}
+	}
+
+	clearIntervals() {
+		for (const [id, t] of Object.entries(this.intervals)) {
+			clearInterval(t)
+			delete this.intervals[parseInt(id, 10)] // stupid ts.
+		}
+	}
+
 }
 
 export async function createContext(iso: ivm.Isolate, bridge: Bridge, opts: ivm.ContextOptions = {}): Promise<Context> {
 	let ctx = new Context(await iso.createContext(opts), iso)
 	await ctx.bootstrap(bridge)
 	return ctx
+}
+
+function tryRelease(rel: Releasable) {
+	try {
+		rel.release()
+	} catch (e) { }
 }