+++ v8env/fly/index.js
@@ -1,8 +1,9 @@
 import flyCacheInit from './cache'
 import flyLogInit from './log'
-import SparkMD5 from 'spark-md5'
 import initStreams from './streams'
 import initImage from '../ts/fly/image.ts'
+import { Document, Element } from '../document'
+
 
 /**
  * @namespace fly
@@ -12,12 +13,9 @@ export default function flyInit(ivm, dispatcher) {
     cache: flyCacheInit(ivm, dispatcher),
     http: require('./http'),
     log: flyLogInit(ivm, dispatcher),
-    util: {
-      md5: {
-        hash: SparkMD5.hash
-      }
-    },
     streams: initStreams(ivm, dispatcher),
-    Image: initImage(ivm, dispatcher)
+    Image: initImage(ivm, dispatcher),
+    //Document,
+    //Element
   }
 }