+++ v8env/ts/fly/image.ts
@@ -1,31 +1,28 @@
 /**
- * The Fly image API.
- * @module fly
- */
-
-/**
  * A class for modifying images. This uses operations from [Sharp](http://sharp.pixelplumbing.com/en/stable/) under the hood.
  * @module fly
  */
+import { transferInto } from "../utils/buffer";
 export class Image {
   /** @hidden */
-  data: ArrayBuffer
-  /** @hidden */
-  operations: Image.Operation[]
+  data: ArrayBuffer | null
   /** @hidden */
   info: Image.Metadata | null
 
+  /** @hidden */
+  private _ref: any
+
   /**
    * Constructs a new Image from raw Buffer data
-   * @param data Raw image data from `fetch` or `cache` or somewhere else.
+   * @param data Raw image data as a buffer, or options to create a new image
    */
-  constructor(data: ArrayBuffer) {
-    if (!(data instanceof ArrayBuffer)) {
-      throw new Error("Data must be an ArrayBuffer")
+  constructor(data: ArrayBuffer | Image.CreateOptions) {
+    this.data = null
+    this._ref = null
+    if (data instanceof ArrayBuffer) {
+      this.data = data
     }
-    //console.log("data:", data.constructor)
-    this.data = data
-    this.operations = []
+    this._ref = constructImage(data)
     this.info = null
   }
 
@@ -36,63 +33,243 @@ export class Image {
    * Pass `undefined` or `null` to auto-scale the width to match the height.
    * @param height Height in pixels of the resulting image. 
    * Pass `undefind` or `null` to auto-scale the height to match the width.
-   * @param options Resize options
-   * @returns {fly.Image}
+   * @param options Resize options}
    */
   resize(width?: number, height?: number, options?: Image.ResizeOptions) {
-    this.operations.push({ name: "resize", args: [width, height, options] })
+    this._imageOperation("resize", width, height, options)
     return this
   }
 
-  crop(...args: any[]) {
-    this.operations.push({ name: "crop", args: args })
+  /**
+   * Overlay (composite) an image over the processed (resized, extracted etc.) image.
+   * 
+   * The overlay image must be the same size or smaller than the processed image. If both top and left options are provided, they take precedence over gravity.
+   *
+   * If the overlay image contains an alpha channel then composition with premultiplication will occur.
+   * @param overlay image to overlay
+   * @param options control how the overlay is composited
+   */
+  overlayWith(overlay: ArrayBuffer | Image, options?: Image.OverlayOptions) {
+    let p: any = overlay
+    if (p instanceof Image) {
+      p = p._ref
+    }
+    this._imageOperation("overlayWith", p, options)
+    return this
+  }
+  /**
+   * Preserving aspect ratio, resize the image to be as large as possible
+   * while ensuring its dimensions are less than or equal to the `width` and `height` specified.
+   *
+   * Both `width` and `height` must be provided via `resize` otherwise the behavior will default to `crop`.
+   *
+   * ```
+   * const data = await new ImageinputBuffer)
+   *   .resize(200, 200)
+   *   .max()
+   *   .toBuffer()
+   * ```
+   */
+  max() {
+    this._imageOperation("max")
     return this
   }
 
-  embed(...args: any[]) {
-    this.operations.push({ name: "embed", args: args })
+  /**
+   * Produce the "negative" of the image.
+   */
+  negate() {
+    this._imageOperation("negate")
     return this
   }
 
-  background(...args: any[]) {
-    this.operations.push({ name: "background", args: args })
+  /**
+   * Crop the resized image to the exact size specified, the default behaviour.
+   *
+   * Possible attributes of the optional `sharp.gravity` are `north`, `northeast`, `east`, `southeast`, `south`,
+   * `southwest`, `west`, `northwest`, `center` and `centre`.
+   *
+   * The experimental strategy-based approach resizes so one dimension is at its target length
+   * then repeatedly ranks edge regions, discarding the edge with the lowest score based on the selected strategy.
+   * - `entropy`: focus on the region with the highest [Shannon entropy](https://en.wikipedia.org/wiki/Entropy_%28information_theory%29).
+   * - `attention`: focus on the region with the highest luminance frequency, colour saturation and presence of skin tones.
+   *
+   * ```
+   * const cropped = await new Image(input)
+   *   .resize(200, 200)
+   *   .crop(Image.strategy.entropy)
+   *   .toBuffer()
+   * ```
+   */
+  crop(crop?: Image.gravity | Image.strategy) {
+    this._imageOperation("crop", crop)
+    return this
+  }
+  /**
+   * Preserving aspect ratio, resize the image to the maximum `width` or `height` specified
+   * then embed on a background of the exact `width` and `height` specified.
+   *
+   * If the background contains an alpha value then WebP and PNG format output images will
+   * contain an alpha channel, even when the input image does not.
+   *
+   * ```
+   * const data = new Image(input)
+   *   .resize(200, 300)
+   *   .background({r: 0, g: 0, b: 0, alpha: 0})
+   *   .embed()
+   *   .toBuffer();
+   * ```
+   * @param gravity embed to an edge/corner, defaults to `center`.
+   */
+  embed(gravity?: Image.gravity) {
+    this._imageOperation("embed", gravity)
+    return this
+  }
+  /**
+   * Set the background for the `embed`, `flatten` and `extend` operations.
+   * The default background is `{r: 0, g: 0, b: 0, alpha: 1}`, black without transparency.
+   *
+   * Delegates to the _color_ module, which can throw an Error
+   * but is liberal in what it accepts, clipping values to sensible min/max.
+   * The alpha value is a float between `0` (transparent) and `1` (opaque).
+   *
+   * @param rgba - parsed by the [color](https://www.npmjs.org/package/color) module to extract values for red, green, blue and alpha.
+   */
+  background(rgba: string | Image.Color) {
+    this._imageOperation("background", rgba)
+    return this
+  }
+  /**
+   * Do not enlarge the output image if the input image width or height are already less than the required dimensions.
+   * This is equivalent to GraphicsMagick's `>` geometry option: _"change the dimensions of the image only if its width
+   * or height exceeds the geometry specification"_. Use with `max()` to preserve the image's aspect ratio
+   * The default behaviour before function call is false, meaning the image will be enlarged.
+   * @param flag Toggle option
+   */
+  withoutEnlargement(flag?: boolean) {
+    this._imageOperation("withoutEnlargement", flag)
     return this
   }
 
-  withoutEnlargement(...args: any[]) {
-    this.operations.push({ name: "withoutEnlargement", args: args })
+  /**
+   * Output image to JPEG
+   * 
+   * ```
+   *  * // Convert any input to very high quality JPEG output
+   * const data = await new Image(input)
+   *   .jpeg({
+   *     quality: 100,
+   *     chromaSubsampling: '4:4:4'
+   *   })
+   *   .toBuffer();
+   * ```
+   * 
+   * @param options JPEG output options
+   */
+  jpeg(options?: Image.JpegOptions) {
+    this._imageOperation("jpeg", options)
+    return this
+  }
+  /**
+   * Use these PNG options for output image.
+   *
+   * PNG output is always full colour at 8 or 16 bits per pixel.
+   * Indexed PNG input at 1, 2 or 4 bits per pixel is converted to 8 bits per pixel.
+   *
+   * ```
+   * // Convert any input to PNG output
+   * const data = await new Image(input)
+   *   .png()
+   *   .toBuffer();
+   * ```
+   * 
+   * @param options Compression and encoding options
+   */
+  png(options?: Image.PngOptions) {
+    this._imageOperation("png", options)
     return this
   }
 
-  png(...args: any[]) {
-    this.operations.push({ name: "png", args: args })
+  /**
+   * Output to webp.
+   *
+   * ```
+   * // Convert any input to lossless WebP output
+   * const data = await new Image(input)
+   *   .webp({ lossless: true })
+   *   .toBuffer();
+   * ```
+   * @param options webp encoding options, quality, etc
+   */
+  webp(options?: Image.WebpOptions) {
+    this._imageOperation("webp", options)
     return this
   }
 
-  webp(...args: any[]) {
-    this.operations.push({ name: "webp", args: args })
+  /**
+   * Include all metadata (EXIF, XMP, IPTC) from the input image in the output image.
+   * The default behaviour, when `withMetadata` is not used, is to strip all metadata
+   * and convert to the device-independent sRGB colour space. This will also convert
+   * to and add a web-friendly sRGB ICC profile.
+   * @param options 
+   *  `options.orientation:` value between 1 and 8, used to update the EXIF
+   * `Orientation` tag.
+   */
+  withMetadata(options?: { orientation: number }) {
+    this._imageOperation("withMedata", options)
     return this
   }
 
-  withMetadata(...args: any[]) {
-    this.operations.push({ name: "withMedata", args: args })
+  /**
+   *  Merge alpha transparency channel, if any, with `background`.
+   * @param flatten Defaults to true
+   */
+  flatten(flatten: boolean) {
+    this._imageOperation("flatten", flatten)
     return this
   }
 
-  async metadata(): Promise<Image.Metadata> {
-    return await imageMetadata(this)
+  /**
+   * Pads image by number of pixels. If image is 200px wide, `extend(20)` makes it 220px wide
+   * @param extend If numeric, pads all sides of an image.
+   * 
+   * Otherwise, pad each side by the specified amount.
+   */
+  extend(extend: number | Image.ExtendOptions) {
+    this._imageOperation("extend", extend)
+    return this
+  }
+
+  /**
+   * Get metadata from image
+   */
+  metadata(): Image.Metadata {
+    const m = imageMetadata(this._ref)
+    this.info = m
+    return m
+  }
+
+  /** @hidden */
+  private _imageOperation(name: string, ...args: any[]) {
+    if (!imageOperation) {
+      throw new Error("Image operations not enabled")
+    }
+    return imageOperation(this._ref, name, ...args)
   }
 
   async toBuffer(): Promise<Image.OperationResult> {
-    if (!modifyImage) {
+    if (!imageToBuffer) {
       throw new Error("Image operations not enabled")
     }
-    const result = await modifyImage(this)
+    const result = await imageToBuffer(this._ref)
     return result
   }
 
   async toImage(): Promise<Image> {
-    const result = await this.toBuffer()
+    if (!imageToBuffer) {
+      throw new Error("Image operations not enabled")
+    }
+    const result = await imageToBuffer(this._ref)
     const i = new Image(result.data)
     i.info = result.info
     return i
@@ -100,10 +277,15 @@ export class Image {
 }
 
 export namespace Image {
-  /** @hidden */
-  export interface Operation {
-    name: string,
-    args: any[]
+  export interface Color {
+    /** red channel, 0-255 */
+    r: number,
+    /** green channel, 0-255 */
+    g: number,
+    /** blue channel, 0-255 */
+    b: number,
+    /** alpha channel, 0.0-1.0 */
+    alpha: number
   }
 
   export interface Metadata {
@@ -131,6 +313,50 @@ export namespace Image {
     data: ArrayBuffer
   }
 
+  /**
+   * WebP output options
+   */
+  export interface WebpOptions {
+    /** quality, integer 1-100, defaults to 80 */
+    quality?: number,
+    /** quality of alpha layer, integer 0-100, default to 100 */
+    alphaQuality?: number,
+    /** use lossless compression mode */
+    lossless?: boolean,
+    /** use near_lossless compression mode */
+    nearLossless?: boolean,
+    /** force WebP output, otherwise attempt to use input format, defaults to true */
+    force?: boolean
+  }
+
+  /**
+   * JPEG output options
+   */
+  export interface JpegOptions {
+    /** quality, integer 1-100, defaults to 80 */
+    quality?: number,
+    /** use progressive (interlace) scan (defaults to false) */
+    progressive?: boolean,
+    /** set to '4:4:4' to prevent chroma subsampling when quality <= 90 */
+    chromaSubsampling?: boolean,
+    /** force output to jpeg (default true) */
+    force?: boolean
+  }
+
+  /**
+   * PNG output options
+   */
+  export interface PngOptions {
+    /** use progressive (interlace) scan (defaults to false) */
+    progressive?: boolean,
+    /** zlib compression level, 0-9 (default to 9) */
+    compressionLevel?: number,
+    /** use adaptive row filtering (defaults to false) */
+    adaptiveFiltering?: boolean,
+    /** force PNG output (defaults to true) */
+    force?: boolean
+  }
+
   export interface ResizeOptions {
     /**
      * the kernel to use for image reduction.
@@ -143,9 +369,33 @@ export namespace Image {
      * some images. (optional, default `true`)
      */
     fastShrinkOnLoad?: boolean
+  }
 
+  export interface OverlayOptions {
+    /**  gravity at which to place the overlay. (optional, default `center`) */
+    gravity?: gravity,
+    top?: number,
+    left?: number,
+    tile?: Boolean,
+    cutout?: Boolean,
+    density?: number
+  }
+
+  export interface ExtendOptions {
+    top?: number,
+    left?: number,
+    bottom?: number,
+    right?: number
+  }
+
+  export interface CreateOptions {
+    width: number,
+    height: number,
+    channels: number,
+    background: Color | string
   }
 }
+
 /** @hidden */
 interface OperationFunction {
   (image: Image): Promise<Image.OperationResult>
@@ -157,21 +407,47 @@ interface MetadataFunction {
 /**
  * @hidden
  */
-let modifyImage: OperationFunction
+let constructImage: (data: ArrayBuffer | Image.CreateOptions, options?: any) => any
 /**
  * @hidden
  */
-let imageMetadata: MetadataFunction
+let imageOperation: (ref: any, name: string, ...args: any[]) => any
+/** @hidden */
+let imageToBuffer: (ref: any) => Promise<Image.OperationResult>
+/** @hidden */
+let imageMetadata: (ref: any) => Image.Metadata
 
-/**
- * @hidden 
- */
+/** @hidden */
 export default function initImage(ivm: any, dispatcher: any) {
-  modifyImage = async function modifyImage(image: Image): Promise<Image.OperationResult> {
+  constructImage = function (data: ArrayBuffer | Image.CreateOptions, options?: any) {
+    const args: any[] = [undefined, undefined]
+    if (data instanceof ArrayBuffer) {
+      args[0] = transferInto(ivm, data)
+    } else if (data instanceof Object) {
+      args[1] = new ivm.ExternalCopy(data).copyInto({ release: true })
+    }
+    return dispatcher.dispatchSync("fly.Image()", ...args)
+  }
+  imageOperation = function (ref: any, name: string, ...args: any[]) {
+    for (let i = 0; i < args.length; i++) {
+      const a = args[i]
+      if (a instanceof ArrayBuffer) {
+        args[i] = transferInto(ivm, a)
+      } else if (typeof a === "object" && !(a instanceof ivm.Reference)) {
+        args[i] = new ivm.ExternalCopy(a).copyInto({ release: true })
+      }
+    }
+    const result = dispatcher.dispatchSync("fly.Image.operation", ref, name, ...args)
+    if (result != ref) {
+      //console.error("Image ref mismatch:", name, ref.typeof, result.typeof)
+      //throw new Error(["image operation failed, result not expected:", ref.typeof, result.typeof].join(" "))
+    }
+    return result
+  }
+  imageToBuffer = async function (ref: any) {
     return new Promise<Image.OperationResult>((resolve, reject) => {
-      dispatcher.dispatch("flyModifyImage",
-        new ivm.ExternalCopy(image.data).copyInto({ release: true }),
-        new ivm.ExternalCopy(image.operations).copyInto({ release: true }),
+      dispatcher.dispatch("fly.Image.toBuffer",
+        ref,
         new ivm.Reference((err: string, data: ArrayBuffer, info: any) => {
           if (err) {
             reject(err)
@@ -182,19 +458,8 @@ export default function initImage(ivm: any, dispatcher: any) {
       )
     })
   }
-  imageMetadata = async function imageMetadata(image: Image): Promise<Image.Metadata> {
-    return new Promise<Image.Metadata>((resolve, reject) => {
-      dispatcher.dispatch("flyImageMetadata",
-        new ivm.ExternalCopy(image.data).copyInto({ release: true }),
-        new ivm.Reference((err: string, metadata: any) => {
-          if (err) {
-            reject(err)
-            return
-          }
-          resolve(metadata)
-        })
-      )
-    })
+  imageMetadata = function imageMetadata(ref): Image.Metadata {
+    return dispatcher.dispatchSync("fly.Image.metadata", ref)
   }
   return Image
 }