+++ v8env/ts/types/dispatcher.d.ts
@@ -1,0 +1,9 @@
+/**
+ * @hidden
+ * @private
+ * @module fly
+ */
+export class Dispatcher {
+  dispatchSync: (name: string, ...args: any[]) => any
+  dispatch: (name: string, ...args: any[]) => Promise<any>
+}